+++ rust/src/lib.rs
@@ -3,4 +3,6 @@ pub mod error;
 pub mod oauth;
 pub mod options;
 pub mod responses;
+
+#[cfg(feature = "webhook")]
 pub mod webhook;